% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_new_entry_to_table.R
\name{append_new_entry_to_table}
\alias{append_new_entry_to_table}
\title{append_new_entry_to_table}
\usage{
append_new_entry_to_table(
  pathogen = NA,
  table_type = NA,
  new_row = NA,
  validate = TRUE,
  write_table = FALSE,
  vignette_prepend = ""
)
}
\arguments{
\item{pathogen}{name of pathogen}

\item{new_row}{all the required details for the new row}

\item{validate}{check if the new row to be added passes minimum criteria}

\item{write_table}{write appended table}
}
\value{
Null (updates table)
}
\description{
append_new_entry_to_table
}
