% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_qa_score.R
\name{assign_qa_score}
\alias{assign_qa_score}
\title{Assign quality assessment score to each article}
\usage{
assign_qa_score(articles, ignore_errors = FALSE)
}
\arguments{
\item{articles}{data.frame loaded from \code{load_epidata} function}
}
\value{
article data.frame with an updated column containing QS score out of
100.
}
\description{
Assign quality assessment score to each article
}
