% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_epidata_raw.R
\name{check_column_types}
\alias{check_column_types}
\title{Checks that the column types of the input csv matches the column types
expected by epireview.}
\usage{
check_column_types(file_path, col_types, raw_colnames)
}
\description{
This function creates a vroom object (the same type created by read_csv) and
checks if there are any problems with the file. If there are it will provide
the offending columns and the number problematic rows (per column).
A csv with the details of the issue will be written to a tmp file and
the location will be provided.
This function will prevent data from being loaded until all column types are
correct.
}
\details{
The function is intended to be used
internally by \code{load_epidata_raw} where the files are being read.
}
\seealso{
article_column_type parameter_column_type, outbreak_column_type, model_column_type
}
