% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_ulim}
\alias{check_ulim}
\title{Check upper limit of parameter values}
\usage{
check_ulim(df, ulim, param)
}
\arguments{
\item{df}{A data frame containing the parameter values.}

\item{ulim}{The specified upper limit for the parameter values.}

\item{param}{The name of the parameter.}
}
\value{
The suggested upper limit for the parameter values.
}
\description{
Check upper limit of parameter values
}
\details{
Each forest plot has a default upper limit for the parameter values, based on
the expected range of the parameter. This function checks the upper limit of
parameter values in a data frame and compares it with the default limit.
If the maximum parameter value
exceeds the specified upper limit, a warning message is displayed. The
function also returns the suggested upper limit. It is used internally by
the forest plot functions (only for warning the user). You can also use it
directly to update the default upper limit in the forest plot functions.
}
\examples{
df <- data.frame(parameter_value = c(10, 20, 30))
check_ulim(df, 25, "parameter")

}
