% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{country_palette}
\alias{country_palette}
\title{country_palette Function}
\usage{
country_palette(x)
}
\arguments{
\item{x}{A vector of country names. If provided, the function will return a color palette for the specified countries.}
}
\value{
A color palette for the specified countries.
}
\description{
This function returns a color palette for countries.
}
\examples{
# Get color palette for all countries
country_palette()

# Get color palette for specific countries
country_palette(c('Liberia', 'Guinea', 'Sierra Leone'))

}
