% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_new_article_entry.R,
%   R/create_new_parameter_entry.R
\name{create_new_article_entry}
\alias{create_new_article_entry}
\title{Create new article entry}
\usage{
create_new_article_entry(
  pathogen = NA,
  new_param = c(list(article_id = as.integer(NA)), list(parameter_type =
    as.character(NA)), list(parameter_value = as.double(NA)), list(parameter_unit =
    as.character(NA)), list(parameter_lower_bound = as.double(NA)),
    list(parameter_upper_bound = as.double(NA)), list(parameter_value_type =
    as.character(NA)), list(parameter_uncertainty_single_value = as.double(NA)),
    list(parameter_uncertainty_singe_type = as.character(NA)),
    list(parameter_uncertainty_lower_value = as.double(NA)),
    list(parameter_uncertainty_upper_value = as.double(NA)), 
    
    list(parameter_uncertainty_type = as.character(NA)), list(cfr_ifr_numerator =
    as.integer(NA)), list(cfr_ifr_denominator = as.integer(NA)), list(distribution_type =
    as.character(NA)), list(distribution_par1_value = as.double(NA)),
    list(distribution_par1_type = as.character(NA)), list(distribution_par1_uncertainty =
    as.logical(NA)), list(distribution_par2_value = as.double(NA)),
    list(distribution_par2_type = as.character(NA)), list(distribution_par2_uncertainty =
    as.logical(NA)), list(method_from_supplement = as.logical(NA)), 
    
    list(method_moment_value = as.character(NA)), list(cfr_ifr_method =
    as.character(NA)), list(method_r = as.character(NA)), list(method_disaggregated_by =
    as.character(NA)), list(method_disaggregated = as.logical(NA)),
    list(method_disaggregated_only = as.logical(NA)), list(riskfactor_outcome =
    as.character(NA)), list(riskfactor_name = as.character(NA)),
    list(riskfactor_occupation = as.character(NA)), list(riskfactor_significant =
    as.character(NA)), list(riskfactor_adjusted = as.character(NA)), 
    
    list(population_sex = as.character(NA)), list(population_sample_type =
    as.character(NA)), list(population_group = as.character(NA)), list(population_age_min
    = as.integer(NA)), list(population_age_max = as.integer(NA)),
    list(population_sample_size = as.integer(NA)), list(population_country =
    as.character(NA)), list(population_location = as.character(NA)),
    list(population_study_start_day = as.integer(NA)), list(population_study_start_month
    = as.character(NA)), list(population_study_start_year = as.integer(NA)), 
    
    list(population_study_end_day = as.integer(NA)), list(population_study_end_month =
    as.character(NA)), list(population_study_end_year = as.integer(NA)), list(genome_site
    = as.character(NA)), list(genomic_sequence_available = as.logical(NA)),
    list(parameter_class = as.character(NA)), list(covidence_id = as.integer(NA)),
    list(Uncertainty = as.character(NA)), list(`Survey year` = as.character(NA)))
)

create_new_article_entry(
  pathogen = NA,
  new_param = c(list(article_id = as.integer(NA)), list(parameter_type =
    as.character(NA)), list(parameter_value = as.double(NA)), list(parameter_unit =
    as.character(NA)), list(parameter_lower_bound = as.double(NA)),
    list(parameter_upper_bound = as.double(NA)), list(parameter_value_type =
    as.character(NA)), list(parameter_uncertainty_single_value = as.double(NA)),
    list(parameter_uncertainty_singe_type = as.character(NA)),
    list(parameter_uncertainty_lower_value = as.double(NA)),
    list(parameter_uncertainty_upper_value = as.double(NA)), 
    
    list(parameter_uncertainty_type = as.character(NA)), list(cfr_ifr_numerator =
    as.integer(NA)), list(cfr_ifr_denominator = as.integer(NA)), list(distribution_type =
    as.character(NA)), list(distribution_par1_value = as.double(NA)),
    list(distribution_par1_type = as.character(NA)), list(distribution_par1_uncertainty =
    as.logical(NA)), list(distribution_par2_value = as.double(NA)),
    list(distribution_par2_type = as.character(NA)), list(distribution_par2_uncertainty =
    as.logical(NA)), list(method_from_supplement = as.logical(NA)), 
    
    list(method_moment_value = as.character(NA)), list(cfr_ifr_method =
    as.character(NA)), list(method_r = as.character(NA)), list(method_disaggregated_by =
    as.character(NA)), list(method_disaggregated = as.logical(NA)),
    list(method_disaggregated_only = as.logical(NA)), list(riskfactor_outcome =
    as.character(NA)), list(riskfactor_name = as.character(NA)),
    list(riskfactor_occupation = as.character(NA)), list(riskfactor_significant =
    as.character(NA)), list(riskfactor_adjusted = as.character(NA)), 
    
    list(population_sex = as.character(NA)), list(population_sample_type =
    as.character(NA)), list(population_group = as.character(NA)), list(population_age_min
    = as.integer(NA)), list(population_age_max = as.integer(NA)),
    list(population_sample_size = as.integer(NA)), list(population_country =
    as.character(NA)), list(population_location = as.character(NA)),
    list(population_study_start_day = as.integer(NA)), list(population_study_start_month
    = as.character(NA)), list(population_study_start_year = as.integer(NA)), 
    
    list(population_study_end_day = as.integer(NA)), list(population_study_end_month =
    as.character(NA)), list(population_study_end_year = as.integer(NA)), list(genome_site
    = as.character(NA)), list(genomic_sequence_available = as.logical(NA)),
    list(parameter_class = as.character(NA)), list(covidence_id = as.integer(NA)),
    list(Uncertainty = as.character(NA)), list(`Survey year` = as.character(NA)))
)
}
\arguments{
\item{pathogen}{name of pathogen}

\item{new_param}{all the required details for the new parameter...}

\item{new_article}{all the required details for the new article}
}
\value{
return data for new row to be added with append_new_entry_to_table function

return data for new row to be added with append_new_entry_to_table function
}
\description{
Create new article entry

Create new parameter entry
}
\examples{
create_new_article_entry('marburg',new_article = c( list( "first_author_first_name" = as.character("Joe")),
                                                    list( "first_author_surname"    = as.character("Blocks")),
                                                    list( "article_title"           = as.character("hello")),
                                                    list( "doi"                     = as.character(NA)),
                                                    list( "journal"                 = as.character("ABC")),
                                                    list( "year_publication"        = as.integer(2000)),
                                                    list( "volume"                  = as.integer(NA)),
                                                    list( "issue"                   = as.integer(NA)),
                                                    list( "page_first"              = as.integer(NA)),
                                                    list( "page_last"               = as.integer(NA)),
                                                    list( "paper_copy_only"         = as.logical(NA)),
                                                    list( "notes"                   = as.character(NA)),
                                                    list( "qa_m1"                   = as.integer(1)),
                                                    list( "qa_m2"                   = as.integer(0)),
                                                    list( "qa_a3"                   = as.integer(NA)),
                                                    list( "qa_a4"                   = as.integer(1)),
                                                    list( "qa_d5"                   = as.integer(0)),
                                                    list( "qa_d6"                   = as.integer(NA)),
                                                    list( "qa_d7"                   = as.integer(1))))
}
