% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_new_model_entry.R
\name{create_new_model_entry}
\alias{create_new_model_entry}
\title{Create new model entry}
\usage{
create_new_model_entry(
  pathogen = NA,
  new_model = list(article_id = as.numeric(NA), model_type = as.character(NA),
    compartmental_type = as.character(NA), stoch_deter = as.character(NA),
    theoretical_model = as.logical(NA), interventions_type = as.character(NA),
    code_available = as.logical(NA), transmission_route = as.character(NA), assumptions =
    as.character(NA), covidence_id = as.integer(NA))
)
}
\arguments{
\item{pathogen}{name of pathogen}

\item{new_model}{all the required details for the new model}
}
\value{
return new row of data to be added to the model data set using the
append_new_entry_to_table() function
}
\description{
Create new model entry
}
