% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_new_outbreak_entry.R
\name{create_new_outbreak_entry}
\alias{create_new_outbreak_entry}
\title{Create new outbreak entry}
\usage{
create_new_outbreak_entry(pathogen, new_outbreak)
}
\arguments{
\item{pathogen}{name of pathogen}

\item{new_outbreak}{all the required details for the new outbreak as a list}
}
\value{
return new row of data to be added to the outbreak data set using
the append_new_entry_to_table() function
}
\description{
Create new outbreak entry
}
\examples{
create_new_outbreak_entry(
  pathogen = "marburg",
  new_outbreak = list(
    article_id           = as.numeric(1),
    outbreak_start_day   = as.numeric(NA),
    outbreak_start_month = as.character(NA),
    outbreak_start_year  = as.numeric(1999),
    outbreak_end_day     = as.numeric(NA),
    outbreak_end_month   = as.character(NA),
    outbreak_date_year   = as.numeric(2001),
    outbreak_duration_months = as.numeric(NA),
    outbreak_size        = as.numeric(2),
    asymptomatic_transmission = as.numeric(0),
    outbreak_country     = as.character("Tanzania"),
    outbreak_location    = as.character(NA),
    cases_confirmed      = as.numeric(NA),
    cases_mode_detection = as.character(NA),
    cases_suspected      = as.integer(NA),
    cases_asymptomatic   = as.integer(NA),
    deaths               = as.integer(2),
    cases_severe_hospitalised = as.integer(NA),
    covidence_id         = as.integer(2059)
  )
)
}
