% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_delay_table.R
\name{data_delay_table}
\alias{data_delay_table}
\title{Process data for delay table}
\usage{
data_delay_table(pathogen, exclude)
}
\arguments{
\item{pathogen}{name of pathogen e.g. "marburg"}

\item{exclude}{article IDs to exclude}
}
\value{
processed data to use as input for delay_table()
}
\description{
Process data for delay table
}
\examples{
df <- data_delay_table(pathogen, exclude = c(17, 15))
}
