% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_forest_plot_delay.R
\name{data_forest_plot_delay}
\alias{data_forest_plot_delay}
\title{Process data for use in forest_plot_delay()}
\usage{
data_forest_plot_delay(pathogen, exclude)
}
\arguments{
\item{pathogen}{name of pathogen e.g. "marburg"}

\item{exclude}{article IDs to exclude}
}
\value{
processed data to use as input for forest_plot_delay()
}
\description{
Process data for use in forest_plot_delay()
}
\examples{
data_forest_plot_delay(pathogen = "marburg", exclude = c(15, 17))
}
