% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_outbreak_table.R
\name{data_outbreak_table}
\alias{data_outbreak_table}
\title{Process data for outbreak table}
\usage{
data_outbreak_table(pathogen, prepend = "")
}
\arguments{
\item{pathogen}{name of pathogen e.g. "marburg"}

\item{prepend}{string to allow loading data in vignettes}
}
\value{
processed data to use as input for outbreak_table()
}
\description{
Process data for outbreak table
}
\examples{
df <- data_outbreak_table(pathogen = "marburg")
}
