% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_risk_table.R
\name{data_risk_table}
\alias{data_risk_table}
\title{Process data for use in risk_table()}
\usage{
data_risk_table(pathogen, exclude)
}
\arguments{
\item{pathogen}{name of pathogen e.g. "marburg"}

\item{exclude}{article IDs to exclude}
}
\value{
processed data to use as input for risk_table()
}
\description{
Process data for use in risk_table()
}
\examples{
data_risk_table(pathogen = "marburg", exclude = c(15, 17))
}
