% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_start.R
\name{date_start}
\alias{date_start}
\title{create a date}
\usage{
date_start(start_day, start_month, start_year)
}
\arguments{
\item{start_day}{start day}

\item{start_month}{start month}

\item{start_year}{start year}
}
\description{
create a date
}
