% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delay_table.R
\name{delay_table}
\alias{delay_table}
\title{Create delay table}
\usage{
delay_table(df, pathogen)
}
\arguments{
\item{df}{data with outbreak information}

\item{pathogen}{name of pathogen}
}
\value{
returns flextable
}
\description{
Create delay table
}
