% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_plot_delay.R
\name{delays_to_days}
\alias{delays_to_days}
\title{This function converts delays in different units (hours, weeks, months) to days.
It checks if all delays are in days and warns the user if not. It then converts
hours to days by dividing by 24, weeks to days by multiplying by 7, and months
to days by multiplying by 30.}
\usage{
delays_to_days(df)
}
\arguments{
\item{df}{A data.frame containing delays and their units. This will typically
be a subset of parameters from the data frame returned by
\code{\link{load_epidata}}.}
}
\value{
Updated data.frame with delays converted to days.
}
\description{
This function converts delays in different units (hours, weeks, months) to days.
It checks if all delays are in days and warns the user if not. It then converts
hours to days by dividing by 24, weeks to days by multiplying by 7, and months
to days by multiplying by 30.
}
\examples{
df <- data.frame(parameter_value = c(24, 7, 1),
                 parameter_unit = c("Hours", "Weeks", "Months"))
delays_to_days(df)
# Output:
#   parameter_value parameter_unit
# 1               1           Days
# 2              49           Days
# 3              30           Days
}
