% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epireview-package.R
\docType{package}
\name{epireview-package}
\alias{epireview}
\alias{epireview-package}
\title{epireview: Tools to update and summarise the latest pathogen data from the Pathogen Epidemiology Review Group (PERG)}
\description{
Contains the latest open access pathogen data from the Pathogen Epidemiology Review Group (PERG). Tools are available to update pathogen databases with new peer-reviewed data as it becomes available, and to summarise the latest data using tables and figures.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mrc-ide.github.io/epireview/}
}

}
\author{
\strong{Maintainer}: Rebecca Nash \email{r.nash@imperial.ac.uk} (\href{https://orcid.org/0000-0002-5213-4364}{ORCID})

Authors:
\itemize{
  \item Christian Morgenstern \email{c.morgenstern@imperial.ac.uk} (\href{https://orcid.org/0000-0003-3735-1130}{ORCID})
  \item Sangeeta Bhatia \email{s.bhatia@imperial.ac.uk} (\href{https://orcid.org/0000-0001-6525-101X}{ORCID})
  \item Richard Sheppard \email{richard.sheppard11@imperial.ac.uk}
  \item Joseph Hicks \email{j.hicks@imperial.ac.uk}
  \item Gina Cuomo-Dannenburg (\href{https://orcid.org/0000-0001-6821-0352}{ORCID})
  \item Ruth McCabe (\href{https://orcid.org/0000-0002-6368-9103}{ORCID})
  \item Kelly McCain (\href{https://orcid.org/0000-0003-2393-2217}{ORCID})
}

}
\keyword{internal}
