% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{epireview_read_file}
\alias{epireview_read_file}
\title{Intended for internal use only; setting desirable defaults for reading in
data files.}
\usage{
epireview_read_file(fname, ...)
}
\arguments{
\item{fname}{The name of the file to be read in.}

\item{...}{Additional arguments to be passed to \code{\link{read_delim}}.}
}
\value{
A data frame read in from the file.
}
\description{
Intended for internal use only; setting desirable defaults for reading in
data files.
}
\details{
This function is intended for internal use only. It is used to read
in data files shipped with the package. The idea is to set desirable defaults
in a single place. Mostly it makes reading files quiet by setting show_col_types
to FALSE.
}
\author{
Sangeeta Bhatia
}
