% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_data_forest_plots.R
\name{filter_cols}
\alias{filter_cols}
\title{Filter columns of a data frame based on specified conditions.}
\usage{
filter_cols(x, cols, funs = c("\%in\%", "==", ">", "<"), vals)
}
\arguments{
\item{x}{A data frame.}

\item{cols}{A character vector specifying the columns to be filtered.}

\item{funs}{A character vector specifying the filter functions for each column.
Each function must be in "\%in", "==", ">", "<".}

\item{vals}{A list of values to be used for filtering columns in \code{cols}.}
}
\value{
A data frame with rows filtered based on the specified conditions.
}
\description{
This function filters the rows of a data frame based on specified conditions for
selected columns.
}
\examples{
x <- fetch_data('marburg')
p <- x$params
filter_cols(p, "parameter_type", "==", "Attack rate")
filter_cols(p, "parameter_type", "\%in\%", list(parameter_type = c("Attack rate", "Seroprevalence - IFA")))


}
