% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_plot.R
\name{forest_plot}
\alias{forest_plot}
\title{Basic forest plot}
\usage{
forest_plot(df)
}
\arguments{
\item{df}{data.frame with the following fields: article, label, mid, low, high
The field 'y' is mapped to the y-axis with 'label' used as a display label.
mid refers to the central estimate. low and high represent the lower and higher ends of the
uncertainty interval}
}
\value{
ggplot2 object
}
\description{
Basic forest plot displays central estimate and uncertainty for a parameter from different studies.
y-axis lists the study labels and the x-axis displays parameter.
}
\author{
Sangeeta Bhatia
}
