% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_plot_delay.R
\name{forest_plot_delay}
\alias{forest_plot_delay}
\title{Create forest plot for human delays}
\usage{
forest_plot_delay(df)
}
\arguments{
\item{df}{processed data with human delay information produced using
data_forest_plots()}
}
\value{
returns plot with a summary of the human delays
}
\description{
Create forest plot for human delays
}
\examples{
df <- data_forest_plots(pathogen = "marburg", exclude = c(15, 17))
forest_plot_delay(df = df)
}
