% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_plot_delay.R
\name{forest_plot_delay_int}
\alias{forest_plot_delay_int}
\title{Create forest plot for human delays}
\usage{
forest_plot_delay_int(df, ulim, reorder_studies, ...)
}
\arguments{
\item{df}{The data frame containing the necessary data for generating the
forest plot.}

\item{ulim}{The upper limit for the x-axis of the plot. Default is 10.}

\item{reorder_studies}{Logical. If TRUE, the studies will be reordered using
the \code{\link{reorder_studies}} function. Default is TRUE.}

\item{...}{Additional arguments to be passed to the
\code{\link[forest_plot]{forest_plot}} function.}
}
\value{
returns plot with a summary of the human delays
}
\description{
Create forest plot for human delays
}
\details{
There are a number of 'delays' that are relevant to the pathogens
we study. Some of the more commonly, and hence likely to be extracted for
most pathogens, are infectious period, incubation period, and
serial interval. However, there are many others reported by relatively few
studies or relevant to only a few pathogens. This function is intended
to serve as a template for creating forest plots for these delays. It will
first reparameterise any delays reported in terms of the gamma distribution
to the mean and standard deviation (see \code{\link{reparam_gamma}}). Then,
any parameters reported as inverse (e.g., per day instead of days) will be
inverted (see \code{\link{invert_inverse_params}}). Finally, all delays will
be converted to days (see \code{\link{delays_to_days}}) before
reordering the studies (if requested) and plotting the forest plot.
We also provide some utility functions for the commonly used delays that are
simply wrapper for this function. If however you are interested in
some other delay, you will need to use this function directly, ensuring that
the data frame you provide has only the relevant delays.
}
\seealso{
\code{\link{forest_plot_serial_interval}}
}
