% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_plot_mutations.R
\name{forest_plot_mutations}
\alias{forest_plot_mutations}
\title{Create forest plot for genetic mutations}
\usage{
forest_plot_mutations(df)
}
\arguments{
\item{df}{data with outbreak information}
}
\value{
returns flextable
}
\description{
Create forest plot for genetic mutations
}
