% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_plot_r.R
\name{forest_plot_r0}
\alias{forest_plot_r0}
\title{forest_plot_r0 function}
\usage{
forest_plot_r0(df, ulim = 10, reorder_studies = TRUE, ...)
}
\arguments{
\item{df}{The data frame containing the necessary data for generating the
forest plot.}

\item{ulim}{The upper limit for the x-axis of the plot. Default is 10.}

\item{reorder_studies}{Logical. If TRUE, the studies will be reordered using
the \code{\link{reorder_studies}} function. Default is TRUE.}

\item{...}{
  Arguments passed on to \code{\link[=forest_plot_rt]{forest_plot_rt}}
  \describe{
    \item{\code{}}{}
  }}
}
\value{
ggplot2 object.
}
\description{
This function generates a forest plot for the reproduction number (Basic R0)
using the provided data frame.
}
\examples{
df <- load_epidata("ebola")[["params"]]
forest_plot_r0(df, ulim = 2.5, reorder_studies = TRUE)

}
