% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_plot_delay.R
\name{forest_plot_serial_interval}
\alias{forest_plot_serial_interval}
\title{Create forest plot for serial interval}
\usage{
forest_plot_serial_interval(df, ulim = 30, reorder_studies = TRUE, ...)
}
\arguments{
\item{df}{The data frame containing the necessary data for generating the
forest plot.}

\item{ulim}{The upper limit for the x-axis of the plot. Default is 10.}

\item{reorder_studies}{Logical. If TRUE, the studies will be reordered using
the \code{\link{reorder_studies}} function. Default is TRUE.}

\item{...}{Additional arguments to be passed to the
\code{\link[forestplot]{forest_plot}} function.}
}
\description{
Create forest plot for serial interval
}
\details{
This function is a wrapper for \code{\link{forest_plot_delay_int}}
that is specifically for the serial interval.
}
