% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_plot_severity.R
\name{forest_plot_severity}
\alias{forest_plot_severity}
\title{Create forest plot for severity}
\usage{
forest_plot_severity(df, outbreak_naive = FALSE)
}
\arguments{
\item{df}{processed data with severity information e.g. IFR and CFR (see
vignette for each pathogen)}
}
\value{
returns plot with a summary of IFR and CFR estimates
}
\description{
Create forest plot for severity
}
\examples{
forest_plot_severity(df = data)
}
