% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_table_field_options.R
\name{get_available_table_field_options}
\alias{get_available_table_field_options}
\title{Utility to obtain available table options to pick from}
\usage{
get_available_table_field_options(
  table_type = NA,
  field = "all",
  vignette_prepend = ""
)
}
\arguments{
\item{table_type}{which type of table (either parameter, outbreak,
model) should be loaded}
}
\value{
return data for table_type and pathogen
}
\description{
Utility to obtain available table options to pick from
}
\examples{
get_available_table_field_options("model",field='Model type')
}
