% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_generic.R
\name{get_generic}
\alias{get_generic}
\title{retrieve all parameters of specified type}
\usage{
get_generic(data, parameter_name)
}
\arguments{
\item{data}{parameter dataframe outputted from load_epidata()}

\item{parameter_name}{name of the parameter type to retrieve, ensuring the name matches that in data}
}
\value{
dataframe with all parameter estimates and specified columns
}
\description{
retrieve all parameters of specified type
}
\examples{
df <- load_epidata(pathogen = "ebola")
get_generic(data = df$params, parameter_name = "Human delay - serial interval")
df <- load_epidata(pathogen = "marburg")
get_generic(data = df$params, parameter_name = "Attack rate")
}
