% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_key_columns.R
\name{get_key_columns}
\alias{get_key_columns}
\title{Subset the epidemiological parameter columns by parameter type}
\usage{
get_key_columns(
  data,
  parameter_name = c("cfr", "delays", "sero", "risk_factors", "reproduction_number",
    "genomic", "attack_rate", "doubling_time", "growth_rate", "overdispersion",
    "relative_contribution"),
  all_columns = FALSE
)
}
\arguments{
\item{data}{The parameter \code{data.frame} (\verb{$param}) from \code{\link[=load_epidata]{load_epidata()}}.}

\item{parameter_name}{A \code{character} string with the parameter name. Options
are: \code{"cfr"}, \code{"delay"}, \code{"sero"}, \code{"risk"},
\code{"reproduction_number"}, and \code{"genomic"}.}

\item{all_columns}{The default is FALSE meaning that only the key columns specified
for the specific #' parameter will be retrieved.If this is TRUE, then all columns
in the dataframe will be retrieved.}
}
\value{
A \code{data.frame} with the key columns for the selected parameter.
}
\description{
Subset the epidemiological parameter columns by parameter type
}
\examples{
lassa_data <- load_epidata("lassa")
lassa_params <- lassa_data$params
cfr_lassa <- get_parameter(
  data = lassa_params,
  parameter_name = "Severity - case fatality rate (CFR)"
)
get_key_columns(data = cfr_lassa, parameter_name = "cfr")
}
