% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_parameter.R
\name{get_specific}
\alias{get_specific}
\alias{get_incubation_period}
\alias{get_serial_interval}
\alias{get_generation_time}
\alias{get_delays}
\alias{get_cfr}
\alias{get_risk_factors}
\alias{get_genomic}
\alias{get_reproduction_number}
\alias{get_seroprevalence}
\alias{get_doubling_time}
\alias{get_attack_rate}
\alias{get_growth_rate}
\alias{get_overdispersion}
\alias{get_relative_contribution}
\title{Retrieve all incubation period parameters for a given pathogen}
\usage{
get_incubation_period(data, ...)

get_serial_interval(data, ...)

get_generation_time(data, ...)

get_delays(data, ...)

get_cfr(data, ...)

get_risk_factors(data, ...)

get_genomic(data, ...)

get_reproduction_number(data, ...)

get_seroprevalence(data, ...)

get_doubling_time(data, ...)

get_attack_rate(data, ...)

get_growth_rate(data, ...)

get_overdispersion(data, ...)

get_relative_contribution(data, ...)
}
\arguments{
\item{data}{parameter dataframe output from \code{\link{load_epidata}}}
}
\value{
dataframe with all parameter estimates of this type and key columns (see \code{\link{get_key_columns}})
}
\description{
Retrieve all incubation period parameters for a given pathogen

Retrieve all serial interval parameters for a given pathogen

Retrieve all generation time parameters for a given pathogen

Retrieve all delay parameters for a given pathogen

Retrieve all CFR parameters for a given pathogen

Retrieve all risk factor parameters for a given pathogen

Retrieve all genomic parameters for a given pathogen

Retrieve all reproduction number parameters for a given pathogen

Retrieve all seroprevalence parameters for a given pathogen

Retrieve all doubling time parameters for a given pathogen

Retrieve all attack rate parameters for a given pathogen

Retrieve all growth rate parameters for a given pathogen

Retrieve all overdispersion parameters for a given pathogen

Retrieve all overdispersion parameters for a given pathogen
}
\examples{
df <- load_epidata(pathogen = "lassa")
get_incubation_period(data = df$params)
}
