% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_plot_delay.R
\name{invert_inverse_params}
\alias{invert_inverse_params}
\title{Inverts the values of selected parameters in a data frame.
Sometimes parameters are reported in inverse form (e.g., a delay might be
reported as per day instead of days). Here we carry out a very simple
transformation to convert these to the correct form by inverting the parameter
value and the uncertainty bounds. This may not be appropriate in all cases,
and must be checked on a case-by-case basis.
This function takes a data frame as input and inverts the values of selected parameters.
The selected parameters are identified by the column 'inverse_param'.
The function performs the following operations:
\itemize{
\item Inverts the parameter values of the selected parameters.
\item Swaps the upper and lower bounds of the selected parameters.
\item Inverts the uncertainty values of the selected parameters.
\item Updates the logical vector to indicate that the parameters are no longer inverted.
\item Does not change the unit of the parameters, as it remains the same as the original parameter.
}}
\usage{
invert_inverse_params(df)
}
\arguments{
\item{df}{A data frame containing the parameters to be inverted.}
}
\value{
The input data frame with the selected parameters inverted.
}
\description{
Inverts the values of selected parameters in a data frame.
Sometimes parameters are reported in inverse form (e.g., a delay might be
reported as per day instead of days). Here we carry out a very simple
transformation to convert these to the correct form by inverting the parameter
value and the uncertainty bounds. This may not be appropriate in all cases,
and must be checked on a case-by-case basis.
This function takes a data frame as input and inverts the values of selected parameters.
The selected parameters are identified by the column 'inverse_param'.
The function performs the following operations:
\itemize{
\item Inverts the parameter values of the selected parameters.
\item Swaps the upper and lower bounds of the selected parameters.
\item Inverts the uncertainty values of the selected parameters.
\item Updates the logical vector to indicate that the parameters are no longer inverted.
\item Does not change the unit of the parameters, as it remains the same as the original parameter.
}
}
\examples{
df <- data.frame(parameter_value = c(2, 3, 4),
                 parameter_upper_bound = c(5, 6, 7),
                 parameter_lower_bound = c(1, 2, 3),
                 parameter_uncertainty_upper_value = c(0.1, 0.2, 0.3),
                 parameter_uncertainty_lower_value = c(0.4, 0.5, 0.6),
                 inverse_param = c(FALSE, TRUE, FALSE))
invert_inverse_params(df)
# Output:
}
