% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_key_columns.R
\name{key_columns}
\alias{key_columns}
\title{Key columns across all parameters in the \verb{$params} table from
\code{\link[=load_epidata]{load_epidata()}}}
\usage{
key_columns()
}
\value{
A \code{character} vector.
}
\description{
Key columns across all parameters in the \verb{$params} table from
\code{\link[=load_epidata]{load_epidata()}}
}
\keyword{internal}
