% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{Load data for a particular pathogen}
\usage{
load_data(table_type = NA, pathogen = NA)
}
\arguments{
\item{table_type}{which type of table (either article, parameter, outbreak,
model) should be loaded}

\item{pathogen}{name of pathogen}
}
\value{
return data for table_type and pathogen
}
\description{
Load data for a particular pathogen
}
\examples{
load_data("outbreak","marburg")
}
