% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_epidata.R
\name{load_epidata}
\alias{load_epidata}
\title{Load data for a particular pathogen}
\usage{
load_epidata(table_type = NA, pathogen = NA, vignette_prepend = "")
}
\arguments{
\item{table_type}{which type of table (either "article", "parameter",
"outbreak", or "model") should be loaded}

\item{pathogen}{name of pathogen}

\item{vignette_prepend}{string to allow loading data in vignettes}
}
\value{
return data for specified table_type and pathogen
}
\description{
Load data for a particular pathogen
}
\examples{
load_epidata(table_type = "outbreak", pathogen = "marburg")
}
