% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{marburg_article}
\alias{marburg_article}
\title{Data on the articles identified and included in the systematic review of
articles related to Marburg virus disease.}
\format{
A csv file containing the following parameters:

XXYY
}
\source{
Cuomo-Dannenburg G, McCain K, McCabe R, Unwin HJT, Doohan P, Nash RK,
et al. Marburg Virus Disease outbreaks, mathematical models, and disease
parameters: a Systematic Review. medRxiv; 2023. 2023.07.10.23292424.
Available from: https://www.medrxiv.org/content/10.1101/2023.07.10.23292424v1
}
\description{
This data set provides the details of all extracted articles included in
the systematic review for Marburg virus disease (MVD).
}
