% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{marburg_dropdown_models}
\alias{marburg_dropdown_models}
\title{Dropdown menu options for model extractions in the systematic review of
articles related to Marburg virus disease (MVD).}
\format{
A csv file containing the dropdown options for:
\itemize{
\item Model type
\item Stochastic or deterministic
\item Transmission route
\item Compartment type
\item Assumptions
\item Interventions
}
}
\source{
Cuomo-Dannenburg G, McCain K, McCabe R, Unwin HJT, Doohan P, Nash RK,
et al. Marburg Virus Disease outbreaks, mathematical models, and disease
parameters: a Systematic Review. medRxiv; 2023. 2023.07.10.23292424.
Available from: https://www.medrxiv.org/content/10.1101/2023.07.10.23292424v1
}
\description{
This data set provides all the dropdown menu options extractors had when
extracting data for mathematical models applied to MVD.
}
