% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{marburg_dropdown_outbreaks}
\alias{marburg_dropdown_outbreaks}
\title{Dropdown menu options for outbreak extractions in the systematic review of
articles related to Marburg virus disease (MVD).}
\format{
A csv file containing the dropdown options for:
\itemize{
\item Outbreak country
\item Detection mode
}
}
\source{
Cuomo-Dannenburg G, McCain K, McCabe R, Unwin HJT, Doohan P, Nash RK,
et al. Marburg Virus Disease outbreaks, mathematical models, and disease
parameters: a Systematic Review. medRxiv; 2023. 2023.07.10.23292424.
Available from: https://www.medrxiv.org/content/10.1101/2023.07.10.23292424v1
}
\description{
This data set provides all the dropdown menu options extractors had when
extracting outbreak data for MVD.
}
