% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{marburg_dropdown_parameters}
\alias{marburg_dropdown_parameters}
\title{Dropdown menu options for parameter extractions in the systematic review of
articles related to Marburg virus disease (MVD).}
\format{
A csv file containing the dropdown options for:
\itemize{
\item Population Country
\item Parameter type
\item Units
\item Parameter uncertainty - single type
\item Parameter uncertainty - paired type
\item Distribution type
\item Distribution parameter 1 - type
\item Distribution parameter 2 - type
\item Disaggregated by
\item Sex
\item Setting
\item Group
\item Timing
\item Reproduction number method
\item Time from
\item Time to
\item IFR_CFR_method
\item Riskfactor outcome
\item Riskfactor name
\item Riskfactor occupation
\item Riskfactor significant
\item Riskfactor adjusted
}
}
\source{
Cuomo-Dannenburg G, McCain K, McCabe R, Unwin HJT, Doohan P, Nash RK,
et al. Marburg Virus Disease outbreaks, mathematical models, and disease
parameters: a Systematic Review. medRxiv; 2023. 2023.07.10.23292424.
Available from: https://www.medrxiv.org/content/10.1101/2023.07.10.23292424v1
}
\description{
This data set provides all the dropdown menu options extractors had when
extracting parameter data for MVD.
}
