% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_epidata.R
\name{outbreak_column_type}
\alias{outbreak_column_type}
\title{outbreak_column_type}
\usage{
outbreak_column_type()
}
\value{
A list of column types for the article data frame
}
\description{
This function defines the column types for the outbreaks in the dataset.
It returns a list of column types with their corresponding names.
}
\examples{
outbreak_column_type()
}
\seealso{
parameter_column_type, outbreak_column_type, model_column_type
}
\keyword{column}
\keyword{dataset,}
\keyword{types}
