% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_utils.R
\name{parameter_palette}
\alias{parameter_palette}
\title{Define a consistent color palette for use in
figures}
\usage{
parameter_palette(x = NULL)
}
\arguments{
\item{x}{a list of parameters. Optional. If missing, the entire
palette is returned.}
}
\value{
a named list of colors that can be used
in forest plots for manually setting colors
with for example
\code{\link[ggplot2:scale_color_manual]{scale_color_manual}}
}
\description{
Define a consistent color palette for use in
figures
}
\examples{
parameter_palette()
}
\author{
Sangeeta Bhatia
}
