% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_data_forest_plots.R
\name{prepare_data_forest_plot}
\alias{prepare_data_forest_plot}
\title{Prepapre pathogen-specific data for input use in forest plots}
\usage{
prepare_data_forest_plot(df, cols, funs, vals)
}
\arguments{
\item{cols}{A character vector specifying the columns to be filtered.}

\item{funs}{A character vector specifying the filter functions for each column.
Each function must be one of "in", "==", ">", "<" in quotes.}

\item{vals}{A list of values to be used for filtering columns in \code{cols}.}
}
\value{
a filtered data.frame with the following columns: article_label, parameter_type,
mid, low, high
}
\description{
Prepapre pathogen-specific data for input use in forest plots
}
\details{
This function prepares the pathogen-specific data for use in forest plots.
It applies the desired filters using \code{\link{filter_cols}} and then
renames the columns to be used in the forest plot. Central value is renamed
to "mid", lower and upper uncertainty bounds are renamed to "low" and "high"
respectively. The resulting data.frame can be used to create a basic forest plot using
\code{\link{forest_plot}}. The basic workflow is (a) load data, (b) filter and prepare data, and
(c) create forest plot.
}
