% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{priority_pathogens}
\alias{priority_pathogens}
\title{priority_pathogens}
\usage{
priority_pathogens()
}
\value{
data.frame with the following fields
\itemize{
\item pathogen: name of the pathogen as used in the package
\item articles_screened: number of titles and abstracts screened for inclusion
\item articles_extracted: number of articles from which data were extracted
\item doi: doi of the accompanying systematic review
}
}
\description{
This data set gives the list of WHO priority pathogens for which
the Pathogen Epidemiology Review Group (PERG) has carried out a
systematic review. The data set gives the name of the pathogen
as used in the package and associated information with the review.
}
\details{
Data on the priority pathogens included in the systematic review
}
\examples{
priority_pathogens()
}
