% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_qa_score.R
\name{qa_questions}
\alias{qa_questions}
\title{Quality assessment questionnaire
This function returns the list of 7 questions used to assess quality of
articles.}
\usage{
qa_questions()
}
\value{
a data.frame with the following columns: (a) qnames: these are the
names of the corresponding columns in the articles data.frame; (b) qtext:
the text of the question, and (c) notes: any additional notes.
}
\description{
Quality assessment questionnaire
This function returns the list of 7 questions used to assess quality of
articles.
}
