% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorder_studies.R
\name{reorder_studies}
\alias{reorder_studies}
\title{Reorder articles based on parameter value}
\usage{
reorder_studies(df)
}
\arguments{
\item{df}{The input dataframe to be reordered}
}
\value{
The reordered dataframe
}
\description{
This function takes a dataframe as input (this will typically be params
data.frame from the output of \code{\link{load_epidata}}) and reorders it
to provide a sensible order for plotting.
For each country, studies are ordered by the parameter value or the midpoint
of the parameter range if the parameter value is missing. It creates a new
column 'y' which is an ordered factor with levels corresponding to the
article_label. To order the studies in some other way, set reorder_studies to
FALSE in the parameter specific
forest plot functions (e.g. \code{\link{forest_plot_rt}}).
}
\examples{
ebola <- load_epidata("ebola")
params <- ebola$params
rt <- filter_cols(
  params, "parameter_type", "in", "Reproduction number (Effective, Re)")
reorder_studies(rt)


}
\seealso{
\code{\link{forest_plot_rt}}  \code{\link{forest_plot_r0}}
}
