% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{reparam_gamma}
\alias{reparam_gamma}
\title{Reparameterize Gamma Distribution}
\usage{
reparam_gamma(df)
}
\arguments{
\item{df}{A data frame with updated columns for parameter value and uncertainty.}
}
\value{
data.frame modified data frame with reparameterized gamma distributions.
}
\description{
This function reparameterizes the gamma distribution in a given data frame.
If a parameter has been expressed as a gamma distribution with shape and scale,
we convert these to mean and standard deviation for plotting.
}
