% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_key_columns.R
\name{reproduction_number_key_columns}
\alias{reproduction_number_key_columns}
\title{Key columns for reproduction number parameters in the \verb{$params} table from
\code{\link[=load_epidata]{load_epidata()}}}
\usage{
reproduction_number_key_columns()
}
\value{
A \code{character} vector.
}
\description{
Key columns for reproduction number parameters in the \verb{$params} table from
\code{\link[=load_epidata]{load_epidata()}}
}
\keyword{internal}
