% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_table.R
\name{risk_table}
\alias{risk_table}
\title{Create risk table}
\usage{
risk_table(df, pathogen, supplement = FALSE)
}
\arguments{
\item{df}{data with outbreak information}

\item{pathogen}{name of pathogen}

\item{supplement}{either TRUE or FALSE. If TRUE, returns supplementary
figures.}
}
\value{
returns flextable
}
\description{
Create risk table
}
\examples{
risk_table(df = data, pathogen = "marburg")
}
