% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sero_table.R
\name{sero_table}
\alias{sero_table}
\title{Create seroprevalence table}
\usage{
sero_table(df, pathogen)
}
\arguments{
\item{df}{processed data with parameter information produced by
data_param_table()}

\item{pathogen}{name of pathogen}
}
\value{
returns flextable with a summary of seroprevalence data
}
\description{
Create seroprevalence table
}
\examples{
df <- data_param_table(pathogen = "marburg", exclude = c(15, 17))
sero_table(df = df, pathogen = "marburg")
}
