% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_utils.R
\name{theme_epireview}
\alias{theme_epireview}
\title{Plotting theme for epireview
A standard theme for figures in epireview.}
\usage{
theme_epireview(
  base_size = 11,
  base_family = "",
  base_line_size = base_size/22,
  base_rect_size = base_size/22
)
}
\arguments{
\item{base_size}{base font size, given in pts.}

\item{base_family}{base font family}

\item{base_line_size}{base size for line elements}

\item{base_rect_size}{base size for rect elements}
}
\description{
Plotting theme for epireview
A standard theme for figures in epireview.
}
