% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_article_label.R
\name{update_article_info}
\alias{update_article_info}
\title{Include key article information when DOI is missing}
\usage{
update_article_info(articles)
}
\arguments{
\item{articles}{A data frame containing information about the articles. This
will typically be the output of \code{load_epidata_raw}.}
}
\value{
A modified data frame with an updated column "doi"
with NA values replaced with either (1) article_label appended to include
journal where available or (2) just article_label when journal entry is NA.
}
\description{
This function appends pretty_article_label to include the journal and we use
this information to fill in any entry where the DOI is missing. (Note: DOIs
were introduced in the late 1990s and so articles from before this time often
do not have one.)
}
\examples{
articles <- data.frame(
  doi = c("10.123", NA, "10.234"),
  article_label = c("Smith 2020", "Smith 1979", "Smith 2023"),
  journal = c("Science", "Nature", "The Lancet")
)
update_article_info(articles)

}
