% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_qa_score.R
\name{assign_qa_score}
\alias{assign_qa_score}
\title{Assign quality assessment score to each article}
\usage{
assign_qa_score(articles, ignore_errors = FALSE)
}
\arguments{
\item{articles}{data.frame loaded from \code{load_epidata} function}

\item{ignore_errors}{logical; if \code{TRUE}, the function will assign QA scores
where possible (i.e. where all answers to quality assessment questions
are not NA) and set the QA score to NAfor articles where all answers are NA.
If \code{FALSE}, an error is thrown instead.}
}
\value{
a named list consisting of two elements. The first element of the
list is the article data.frame with an updated column containing three new columns:
\code{qs_denominator} (total number of questions answered), \code{qs_numerator}
(number of questions answered 'yes') and \code{qa_score} (QA score). The second
element of the list (named errors) is a data.frame containing articles
with all NA answers.
}
\description{
Assign quality assessment score to each article
}
\details{
We have used a bespoke 7 question quality assessment (QA) questionnaire to
assess the quality of articles. The questions can be retrieved using the
\code{qa_questions} function. The function assigns a QA score to each article
as the number of questions answered 'yes' divided by the total number of
questions answered (an answer might be NA if the question is not relevant to
the article under consideration).
Articles with all NA answers are excluded from the QA unless \code{ignore_errors}
is set to \code{TRUE}.
}
\examples{
lassa <- load_epidata("lassa")
lassa_qa <- assign_qa_score(lassa$articles, ignore_errors = FALSE)
head(lassa_qa$articles[, c("qa_denominator", "qa_numerator", "qa_score")])
}
\seealso{
\code{\link{qa_questions}}
}
