% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{color_palette}
\alias{color_palette}
\title{Define a consistent color palette for use in
figures. Palettes are currently defined for
parameters and countries. Any other variable will
return NULL}
\usage{
color_palette(col_by = c("parameter_type", "population_country"), ...)
}
\description{
Define a consistent color palette for use in
figures. Palettes are currently defined for
parameters and countries. Any other variable will
return NULL
}
