% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{country_palette}
\alias{country_palette}
\title{country_palette Function}
\usage{
country_palette(countries)
}
\arguments{
\item{countries}{A vector of country names. If not provided, a palette of length 36 with a default set of
countries is returned.}
}
\value{
A named color palette.
}
\description{
This function creates a color palette for countries.
}
\examples{
country_palette()
country_palette(countries = c("USA", "Canada", "Mexico"))

}
