% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_new_article_entry.R
\name{create_new_article_entry}
\alias{create_new_article_entry}
\title{Create new article entry}
\usage{
create_new_article_entry(
  pathogen = NA,
  new_article = list(first_author_first_name = as.character(NA), first_author_surname =
    as.character(NA), article_title = as.character(NA), doi = as.character(NA), journal =
    as.character(NA), year_publication = as.integer(NA), volume = as.integer(NA), issue =
    as.integer(NA), page_first = as.integer(NA), page_last = as.integer(NA),
    paper_copy_only = as.logical(NA), notes = as.character(NA), qa_m1 = as.integer(NA),
    qa_m2 = as.integer(NA), qa_a3 = as.integer(NA), qa_a4 = as.integer(NA), qa_d5 =
    as.integer(NA), qa_d6 = as.integer(NA), 
     qa_d7 = as.integer(NA))
)
}
\arguments{
\item{pathogen}{name of pathogen}

\item{new_article}{all the required details for the new article}
}
\value{
return new row of data to be added to the article data set using the
append_new_entry_to_table() function
}
\description{
Create new article entry
}
