% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_new_model_entry.R
\name{create_new_model_entry}
\alias{create_new_model_entry}
\title{Create new model entry}
\usage{
create_new_model_entry(
  pathogen = NA,
  new_model = c(list(article_id = as.integer(NA)), list(model_type = as.character(NA)),
    list(compartmental_type = as.character(NA)), list(stoch_deter = as.character(NA)),
    list(theoretical_model = as.logical(NA)), list(interventions_type =
    as.character(NA)), list(code_available = as.logical(NA)), list(transmission_route =
    as.character(NA)), list(assumptions = as.character(NA)), list(covidence_id =
    as.integer(NA)))
)
}
\arguments{
\item{pathogen}{name of pathogen}

\item{new_model}{all the required details for the new outbreak}

\item{vignette_prepend}{string to allowing loading data from vignettes}
}
\value{
return data for new row to be added with append_new_entry_to_table function
}
\description{
Create new model entry
}
\examples{
create_new_model_entry('marburg',c( list( "article_id"           = as.integer(1)),
                                       list( "model_type"           = as.character("Compartmental")),
                                       list( "compartmental_type"   = as.character("SEIR,SIR")),
                                       list( "stoch_deter"          = as.character("Deterministic")),
                                       list( "theoretical_model"    = as.logical(FALSE)),
                                       list( "interventions_type"   = as.character("Vaccination")),
                                       list( "code_available"       = as.logical(TRUE)),
                                       list( "transmission_route"   = as.character("Sexual")),
                                       list( "assumptions"          = as.character("Unspecified")),
                                       list( "covidence_id"         = as.integer(2059))),
                         vignette_prepend = "")

}
