% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_new_outbreak_entry.R
\name{create_new_outbreak_entry}
\alias{create_new_outbreak_entry}
\title{Create new outbreak entry}
\usage{
create_new_outbreak_entry(pathogen, new_outbreak)
}
\arguments{
\item{pathogen}{name of pathogen}

\item{new_outbreak}{all the required details for the new outbreak as a list}
}
\value{
return new row of data to be added to the outbreak data set using
the append_new_entry_to_table() function
}
\description{
Create new outbreak entry
}
