% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_forest_plots.R
\name{data_forest_plots}
\alias{data_forest_plots}
\title{Process data for use in forest_plot_delay(), forest_plot_mutations(),
forest_plot_r() and forest_plot_severity()}
\usage{
data_forest_plots(pathogen, prepend = "", exclude = NA)
}
\arguments{
\item{pathogen}{name of pathogen e.g. "marburg"}

\item{prepend}{string to allow loading data in vignettes}

\item{exclude}{article IDs to exclude}
}
\value{
processed data to use as input for forest_plot_delay(),
forest_plot_mutations(), forest_plot_r() and forest_plot_severity()
}
\description{
Process data for use in forest_plot_delay(), forest_plot_mutations(),
forest_plot_r() and forest_plot_severity()
}
\examples{
data_forest_plots(pathogen = "marburg", exclude = c(15, 17))
}
