% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delay_table.R
\name{delay_table}
\alias{delay_table}
\title{Create table for human delay parameters}
\usage{
delay_table(df, pathogen)
}
\arguments{
\item{df}{processed data with parameter information produced using
data_delay_table()}

\item{pathogen}{name of pathogen}
}
\value{
flextable with an overview of the delay parameter estimates
extracted from the included studies
}
\description{
Create table for human delay parameters
}
\examples{
# Process data
df <- data_delay_table(pathogen = "marburg", exclude = c(17, 15))

# Get delay table
delay_table(df = df, pathogen = "marburg")
}
