% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_df_for_metamean.R
\name{filter_df_for_metamean}
\alias{filter_df_for_metamean}
\title{Prepare parameter dataframe for meta analysis of means}
\usage{
filter_df_for_metamean(df)
}
\arguments{
\item{df}{a parameter dataframe. This must have columns for each of the
following: parameter_value, parameter_unit, population_sample_size,
parameter_value_type, parameter_uncertainty_singe_type,
parameter_uncertainty_type, parameter_uncertainty_lower_value,
parameter_uncertainty_upper_value}
}
\value{
a parameter dataframe with relevant rows selected
}
\description{
Prepare parameter dataframe for meta analysis of means
}
\details{
The function checks that the format of df is adequate for conducting
a meta analysis of means, returns errors where this is not the case,
and filters the rows which do not match the required format.
}
\examples{
## TODO: rewrite example once the curation functions exist with examples
## of how to then use metamean function
df <- readRDS("tests/testthat/data/example_param_df.RDS")
dataframe <- filter_df_for_metamean(df)
## mtan <- metamean(data = dataframe, ...)

}
