% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_df_for_metaprop.R
\name{filter_df_for_metaprop}
\alias{filter_df_for_metaprop}
\title{Prepare parameter dataframe for meta analysis of proportions}
\usage{
filter_df_for_metaprop(df, num_col, denom_col)
}
\arguments{
\item{df}{a parameter dataframe. This must have columns for each of the
following: parameter_value, parameter_unit, plus two columns for the
numerator and the denominator of the proportion of interest.
This dataframe will typically be the \code{params}
data.frame from the output of \code{\link{load_epidata}}.}

\item{num_col}{a string specifying the column name for the column containing
the numerator of the proportion of interest.}

\item{denom_col}{a string specifying the column name for the column
containing the denominator of the proportion of interest.}
}
\value{
a parameter dataframe with relevant rows selected to enable
meta analysis of proportions.
}
\description{
Prepare parameter dataframe for meta analysis of proportions
}
\details{
The function checks that the format of df is adequate for conducting a meta
analysis of proportions. It filters the dataframe to only include rows that
meet the required format by (1) removing rows where the denominator is missing,
and (2) removing rows where both the numerator column or parameter value are
missing. If the numerator column is missing and the parameter value is present,
the numerator is imputed as the parameter value divided by 100 times the
denominator.
}
\examples{
## preparing data for meta analyses of CFR for Lassa

df <- load_epidata("lassa")[["params"]]
cfr_df <- df[df$parameter_type \%in\% "Severity - case fatality rate (CFR)",]
cfr_filtered <- filter_df_for_metaprop(cfr_df,
  num_col = "cfr_ifr_numerator", denom_col = "cfr_ifr_denominator")
## cfr_filtered could then be used directly in meta analyses as:
## mtan <- metaprop(data = cfr_filtered, ...)
}
