% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_plot.R
\name{forest_plot}
\alias{forest_plot}
\title{Basic forest plot}
\usage{
forest_plot(
  df,
  facet_by = NA,
  shape_by = NA,
  col_by = NA,
  shp_palette = NA,
  col_palette = NA
)
}
\arguments{
\item{df}{The data frame containing the data for the forest plot. data.frame with the following fields: article, label, mid, low, high
The field 'y' is mapped to the y-axis with 'article_label' used as a display label.
mid refers to the central estimate. low and high represent the lower and higher ends of the
uncertainty interval}

\item{facet_by}{(Optional) Variable to facet the plot by.}

\item{shape_by}{(Optional) Variable to shape the points by.}

\item{col_by}{(Optional) Variable to color the points by.}

\item{shp_palette}{(Optional) Palette for shaping the points. Optional unless shape_by is
not one of 'parameter_value_type'.}

\item{col_palette}{Palette for coloring the points. Optional unless col_by is
not one of 'parameter_type' or 'population_country'.}
}
\value{
A ggplot2 object representing the forest plot.
}
\description{
Basic forest plot displays central estimate and uncertainty for a parameter from different studies.
y-axis lists the study labels and the x-axis displays parameter.
}
\details{
Generates a forest plot.

This function generates a forest plot using the provided data frame.

epireview provides a default palette for parameters and countries.
If you wish to color by a different variable, you must provide a palette.
}
\examples{
df <- data.frame(
  mid = c(1, 2, 3),
  article_label = c("A", "B", "C"),
  low = c(0.5, 1.5, 2.5),
  high = c(1.5, 2.5, 3.5),
  uncertainty_type = c("Range", "Range", "Range**")
)
forest_plot(df)
}
