% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_plot_delays.R
\name{forest_plot_delay}
\alias{forest_plot_delay}
\title{Create forest plot for delays}
\usage{
forest_plot_delay(df)
}
\arguments{
\item{df}{data with outbreak information}
}
\value{
returns flextable
}
\description{
Create forest plot for delays
}
