% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_plot_delays.R
\name{forest_plot_delay}
\alias{forest_plot_delay}
\title{Create forest plot for delays}
\usage{
forest_plot_delay(df)
}
\arguments{
\item{df}{processed data with parameter information (see vignette for each
pathogen)}
}
\value{
returns flextable with a summary of the human delays
}
\description{
Create forest plot for delays
}
