% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_plot_other.R
\name{forest_plot_doubling_time}
\alias{forest_plot_doubling_time}
\title{Calculate and plot the forest plot for doubling time}
\usage{
forest_plot_doubling_time(df, ulim = 30, reorder_studies = TRUE, ...)
}
\arguments{
\item{df}{The data frame containing the necessary data for generating the
forest plot.}

\item{ulim}{The upper limit for the x-axis of the plot. Default is 10.}

\item{reorder_studies}{Logical. If TRUE, the studies will be reordered using
the \code{\link{reorder_studies}} function. Default is TRUE.}

\item{...}{Additional arguments to be passed to the
\code{\link[forestplot]{forest_plot}} function.}
}
\value{
A forest plot object.
}
\description{
This function calculates the doubling time and creates a forest plot
to visualize the results.
}
\examples{
df <- data.frame(parameter_type = c("Doubling time", "Doubling time"),
                 value = c(5, 10))
forest_plot_doubling_time(df, ulim = 20, reorder_studies = TRUE)

}
