% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_plot_mutations.R
\name{forest_plot_mutations}
\alias{forest_plot_mutations}
\title{Create forest plot for genetic mutations}
\usage{
forest_plot_mutations(df)
}
\arguments{
\item{df}{processed data with parameter information produced using
data_forest_plots()}
}
\value{
returns plot with a summary of genetic mutations
}
\description{
Create forest plot for genetic mutations
}
\examples{
df = data_forest_plots(pathogen = "marburg", exclude = c(15, 17))
forest_plot_mutations(df = df)
}
