% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_plot_r.R
\name{forest_plot_r}
\alias{forest_plot_r}
\title{Create forest plot for reproduction numbers}
\usage{
forest_plot_r(df)
}
\arguments{
\item{df}{processed data with parameter information produced using
data_forest_plots()}
}
\value{
returns plot with a summary of reproduction number estimates
}
\description{
Create forest plot for reproduction numbers
}
\examples{
df = data_forest_plots(pathogen = "marburg", exclude = c(15, 17))
forest_plot_r(df = df)
}
