% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_plot_r.R
\name{forest_plot_R}
\alias{forest_plot_R}
\title{Create forest plot for reproduction numbers}
\usage{
forest_plot_R(df)
}
\arguments{
\item{df}{processed data with parameter information (see vignette for each
pathogen)}
}
\value{
returns plot with a summary of reproduction number estimates
}
\description{
Create forest plot for reproduction numbers
}
\examples{
forest_plot_R(df = data)
}
