% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_plot_severity.R
\name{forest_plot_severity}
\alias{forest_plot_severity}
\title{Create forest plot for severity}
\usage{
forest_plot_severity(df, outbreak_naive = FALSE)
}
\arguments{
\item{df}{processed data with severity information produced using
data_forest_plots()}

\item{outbreak_naive}{TRUE or FALSE. FALSE by default.}
}
\value{
returns plot with a summary of IFR and CFR estimates
}
\description{
Create forest plot for severity
}
\examples{
df = data_forest_plots(pathogen = "marburg", exclude = c(15, 17))
forest_plot_severity(df = df)
}
