% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_table_field_options.R
\name{get_table_field_options}
\alias{get_table_field_options}
\title{Utility to obtain available fields in a specified table}
\usage{
get_table_field_options(
  pathogen,
  table = c("model", "parameter", "outbreak", "article"),
  field = "all"
)
}
\arguments{
\item{pathogen}{name of pathogen. Must be one of the priority pathogens
exactly as specified in the package. You can get a list of the
priority pathogens currently included in the package by calling}

\item{table}{the table to be loaded. Must be one of
"article", "parameter", "outbreak", or "model"}

\item{field}{can either be "all" to return all fields in the specified
table_type, or a specific variable name that the user wants to return}
}
\value{
data for specified table_type and field
}
\description{
Note: This will need to be extended when other pathogens are added
}
\examples{
get_table_field_options(
                       pathogen = "marburg", "model",
                       field = "all")

get_table_field_options(pathogen = "marburg", "model",
                        field = "Model type")

get_table_field_options(pathogen = "marburg", "parameter",
                        field = "Reproduction number method")
}
