% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_epidata.R
\name{load_epidata}
\alias{load_epidata}
\title{Retrieve pathogen-specific data}
\usage{
load_epidata(pathogen, mark_multiple = TRUE)
}
\arguments{
\item{pathogen}{name of pathogen. Must be one of the priority pathogens
You can get a list of the
priority pathogens currently included in the package by calling the function
\code{\link{priority_pathogens}}.}

\item{mark_multiple}{logical. If TRUE, multiple studies from the same
author in the same year will be marked with a suffix to distinguish them.}
}
\value{
a list of length 4. The first element is a data.frame called "articles"
which contains all of the information about the articles extracted for this
pathogen. The second element is a data.frame called "params" with articles
information (authors, publication year, doi) combined with the parameters.
The third element is a data.frame called "models" with all transmission
models extracted for this pathogen including articles information as above.
The fourth element is a data.frame called "outbreaks" which contains all
of the outbreaks extracted for this pathogen, where available.
}
\description{
Retrieve pathogen-specific data
}
\details{
The data extracted in the systematic review has been stored in three
files - one each for articles, parameters, and transmission models.
Data in the three files can be linked using article identifier.
This function will read in the pathogen-specific articles and
parameters files and join them into a data.frame. This function also
creates user-friendly short labels for the "parameter_type" column in params
data.frame. See \code{\link{short_parameter_type}} for more details.
}
