% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_data_forest_plots.R
\name{load_epidata}
\alias{load_epidata}
\title{Retrieve pathogen-specific data}
\usage{
load_epidata(pathogen)
}
\arguments{
\item{pathogen}{name of pathogen. Must be one of the priority pathogens
exactly as specified in the package. You can get a list of the
priority pathogens currently included in the package by calling
\code{priority_pathogens()}.}
}
\value{
a list of length 2. The first element is a data.frame
called "params" with articles information (authors, publication year)
combined with the parameters. The second element is a data.frame
called "models" with all transmission models extracted for this
pathogen.
}
\description{
Retrieve pathogen-specific data
}
\details{
The data extracted in the systematic review has been stored in three
files - one each for articles, parameters, and transmission models.
Data in the three files can be linked using article identifier.
This function will read in the pathogen-specific articles and
parameters files and join them into a data.frame. The resulting
data set can be used to create a forest plot.
}
