% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{marburg_outbreak.csv}
\alias{marburg_outbreak.csv}
\title{Data on the outbreaks identified in the systematic review of Marburg articles}
\format{
A csv file containing the following parameters:

XXYY
}
\source{
Cuomo-Dannenburg G, McCain K, McCabe R, Unwin HJT, Doohan P, Nash RK,
et al. Marburg Virus Disease outbreaks, mathematical models, and disease
parameters: a Systematic Review \link{Internet}. medRxiv; 2023 \link{cited 2023 Jul 13}.
p. 2023.07.10.23292424.
Available from: https://www.medrxiv.org/content/10.1101/2023.07.10.23292424v1
}
\description{
This data set gives:

XXYY
}
