% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark_multiple_estimates.R
\name{mark_multiple_estimates}
\alias{mark_multiple_estimates}
\title{Distinguish multiple estimates from the same study}
\usage{
mark_multiple_estimates(
  df,
  col = "parameter_type",
  label_type = c("letters", "numbers")
)
}
\arguments{
\item{df}{The data frame containing the estimates.}

\item{col}{The column name for the table type. For parameters this is
"parameter_type"; for models this is "model_type"; for outbreaks this is
"outbreak_country".}

\item{label_type}{Type of labels to add to distinguish multiple estimates.
Must be one of "letters" or "numbers".}
}
\value{
The modified data frame with updated article_label
}
\description{
Distinguish multiple estimates from the same study
}
\details{
If a study has more than one estimate for the same parameter_type/model/outbreak,
we add a suffix to the article_label to distinguish them
otherwise they will be plotted on the same line in the forest plot. Say
we have two estimates for the same parameter_type (p) from the same study (s),
they will then be labeled as s 1 and s 2.
}
\examples{
df <- data.frame(article_label = c("A", "A", "B", "B", "C"),
                 parameter_type = c("X", "X", "Y", "Y", "Z"))
mark_multiple_estimates(df, label_type = "numbers")

}
