% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outbreak_table.R
\name{outbreak_table}
\alias{outbreak_table}
\title{Create outbreak table}
\usage{
outbreak_table(df, pathogen)
}
\arguments{
\item{df}{data with outbreak information}

\item{pathogen}{name of pathogen}
}
\value{
returns flextable
}
\description{
Create outbreak table
}
