% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{param_pm_uncertainty}
\alias{param_pm_uncertainty}
\title{Update parameter uncertainty columns in a data frame}
\usage{
param_pm_uncertainty(df)
}
\arguments{
\item{df}{A data frame containing the parameter uncertainty columns.
This will typically be the output of \code{\link{load_epidata}}.}
}
\value{
The updated data frame with parameter uncertainty columns
}
\description{
This function updates the parameter uncertainty columns in a data frame
when the uncertainty is given by a single value (standard deviation or
standard error). It creates new columns called \verb{low' (parameter central value - uncertainty) and }high' (parameter central value + uncertainty).
These columns are used by \code{\link{forest_plot}} to plot the
uncertainty intervals. If the uncertainty is given by a range, the
midpoint of the range is used as the central value (mid) and the lower and
upper bounds are used as the low and high values respectively.
}
\examples{
df <- data.frame(
  parameter_value = c(10, 20, 30),
  parameter_uncertainty_single_value = c(1, 2, 3),
  parameter_uncertainty_lower_value = c(5, 15, 25),
  parameter_uncertainty_upper_value = c(15, 25, 35),
  parameter_uncertainty_type = c(NA, NA, NA),
  parameter_uncertainty_singe_type = c("Standard Deviation", "Standard Error", NA)
)
updated_df <- param_pm_uncertainty(df)
updated_df

}
