% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_assessment.R
\name{quality_assessment_plots}
\alias{quality_assessment_plots}
\title{Plots for the quality assessment of the included studies}
\usage{
quality_assessment_plots(pathogen = NA, vignette = NA)
}
\arguments{
\item{pathogen}{pathogen data set to consider}

\item{vignette}{if TRUE file path changes so that data can be accessed}
}
\value{
two plots for quality assessment
}
\description{
Plots for the quality assessment of the included studies
}
\examples{
quality_assessment_plots(pathogen = "marburg")
}
