% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_table.R
\name{risk_table}
\alias{risk_table}
\title{Create risk table}
\usage{
risk_table(df, pathogen)
}
\arguments{
\item{df}{data with outbreak information}

\item{pathogen}{name of pathogen}
}
\value{
returns flextable
}
\description{
Create risk table
}
