% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_table.R
\name{risk_table}
\alias{risk_table}
\title{Create risk table}
\usage{
risk_table(df, pathogen, supplement = FALSE)
}
\arguments{
\item{df}{processed data with risk factor information produced by
data_risk_table()}

\item{pathogen}{name of pathogen}

\item{supplement}{either TRUE or FALSE. If TRUE, returns supplementary
figures.}
}
\value{
returns flextable
}
\description{
Create risk table
}
\examples{
df <- data_risk_table(pathogen = "marburg", exclude = c(15, 17))
risk_table(df = df, pathogen = "marburg")
}
