% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seroprevalence_table.R
\name{sero_table}
\alias{sero_table}
\title{Create seroprevalence table}
\usage{
sero_table(df, pathogen)
}
\arguments{
\item{df}{processed data with seroprevalence information (see vignette for
each pathogen)}

\item{pathogen}{name of pathogen}
}
\value{
returns flextable with a summary of seroprevalence data
}
\description{
Create seroprevalence table
}
\examples{
sero_table(df = data, pathogen = "marburg")
}
