% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seroprevalence_table.R
\name{sero_table}
\alias{sero_table}
\title{Create seroprevalence table}
\usage{
sero_table(df, pathogen)
}
\arguments{
\item{df}{data with outbreak information}

\item{pathogen}{name of pathogen}
}
\value{
returns flextable
}
\description{
Create seroprevalence table
}
