% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_utils.R
\name{value_type_palette}
\alias{value_type_palette}
\title{Define a consistent shape palette for use in forest plots}
\usage{
value_type_palette(x = NULL)
}
\arguments{
\item{x}{a list of parameters}
}
\value{
a named list of shapes where names are value types (mean,
median, std dev etc.)
}
\description{
We map shape aesthetic to value type i.e., mean, median etc.
This function defines a shape palette that can be used in forest
plots
}
\author{
Sangeeta Bhatia
}
